bus: action uuid search based on signature (#30)
An object can have several methods or signature with the same name. This
is the case of the ALMotion.moveTo API. In such case we need to search
an action based on the name and the signature.

Unfortunatly QiLoop have concept of compatible signatures. Here is an
example of missmatch:

    (dlv) p sig
    "(Is)<serviceRemoved,serviceID,name>"
    (dlv) p m.Signals
    map[uint32]github.com/lugu/qiloop/type/object.MetaSignal [
	    86: {
		    Uid: 86,
		    Name: "traceObject",
		    Signature: "((IiIm(ll)<timeval,tv_sec,tv_usec>llII)<EventTrace,id,kind,slotI...+75 more",},
	    106: {
		    Uid: 106,
		    Name: "serviceAdded",
		    Signature: "(Is)",},
	    107: {
		    Uid: 107,
		    Name: "serviceRemoved",
		    Signature: "(Is)",},
    ]

Until a proper signature comparison method is found, we rely on the
hack:

For methods, find the first with:
        - name, param and return type
        - name, param type
        - name

For signals and properties:
        - try to match both name and signature
        - if not a tuple, convert to tuple and try again
        - match only the name

This is clearly not satisfactory and a proper signature comparison
algorithm is needed. It can be cased on a flatten conversion to tuple.
Like:
- (ii)<a,b,c> => (ii)
- (ii) => (ii)
- i => (i)