type/encoding: serialization interface
Currently the serialization is not abstracted: generated proxy and stub simply
read and write following the QiMessaging encoding scheme. This works as long as:

- the encoding scheme is known at compile time.

- the two ends expect exactly the same format (no conversion).

Unfortunaly this is not always true:

- the negociated capability map can alter how objects are serialized (ex:
  ObjectPtrUID and MetaObjectCache).

- optional fields, forward and backward requires runtime type conversion.

This package is the fundation to solve those two problems: it abstracts how to
encode and decode values leaving space for states and runtime adjustments. The
proposed interface use reflection to encode/decode types unknown at compile
time.

The performances suffer from this change: this package is two times slower to
read and three times slower to write. The loss of performance is balanced by
those facts:

- the new encoding is still fast since it compares with gob encoding.
- this change is mandatory for completness of compatibility.

This package offers the possiblity to test other encoding scheme. This
can be useful to implement dynamic tools like `qicli call` which takes
JSON parameters.

Here is a benchmark of the solutions:

- BaseLine is the current implementation
- CodeGen use the new scheme without reflection for struct
- Reflect only use reflections
- JSON uses the JSON encoder/decoder
- Gob uses Go binary encoder

    goos: linux
    goarch: amd64
    pkg: github.com/lugu/qiloop/type/encoding
    BenchmarkReadInfoBaseLine-8      5946895               169 ns/op
    BenchmarkReadInfoCodeGen-8       3542067               291 ns/op
    BenchmarkReadInfoReflect-8       4245644               289 ns/op
    BenchmarkReadInfoGob-8           4414666               271 ns/op
    BenchmarkReadInfoJSON-8           629935              1986 ns/op
    BenchmarkWriteInfoBaseLine-8     7648267               160 ns/op
    BenchmarkWriteInfoCodeGen-8      1968880               524 ns/op
    BenchmarkWriteInfoReflect-8      2171401               542 ns/op
    BenchmarkWriteJSON-8              721453              1524 ns/op
    BenchmarkWriteGob-8              1671292               686 ns/op
    PASS
    ok      github.com/lugu/qiloop/type/encoding    18.195s